Updates and Refactors: Changed scaffold filestructure; Added tests; Improved sys classes; Added util/lib classes

+++ system/log/logger.php
@@ -2,16 +2,16 @@
 /**
  * Logger class file.
  *
- * @author Z. Alem <info@alemmedia.com>
+ * @author Z. Alem <info@alemcode.com>
  */
 
-/** 
+/**
  * The Logger singleton records data by assignment to the $record array
  * and displays it at the end of the applications execution.
  *
- * It also displays all session variables. 
- * 
- * It's display level is determined by DEBUG_LEVEL in the main configuration file (config/main.php)
+ * It also displays all session variables.
+ *
+ * It's display level is determined by DEBUG_LEVEL in the main configuration file(config/main.php)
  * Where 0 is no display, 1 is username 'admin' only, and 2 means it is visible to everyone.
  */
 class Logger
@@ -61,8 +61,9 @@ class Logger
 		'warn'		=> false,
 		'info'		=> false,
 		'debug'		=> false,
+		'display_debug'	=> false,
 		'max_log_size'	=> 50000,
-		'max_old_logs'	=> 2,
+		'max_old_logs'	=> 3,
 	);
 
 
@@ -71,18 +72,21 @@ class Logger
 	 */
 	private function __construct()
 	{
-		$this -> rotate();
 	}
 
+	public function __destruct()
+	{
+		$this->rotate();
+	}
 
 	/**
 	 * open - Creates singleton instance of Logger
 	 *
 	 * @return object
 	 */
-	public static function open() 
+	public static function open()
 	{
-		if ( !isset(self::$instance) )
+		if( !isset(self::$instance) )
 			self::$instance = new self();
 		return self::$instance;
 	}
@@ -95,7 +99,7 @@ class Logger
 	 */
 	public static function setStartTime( $start_time )
 	{
-		self::open() -> start_time = $start_time; 
+		self::open()->start_time = $start_time;
 	}
 
 	/**
@@ -105,7 +109,7 @@ class Logger
 	 */
 	public static function setLogFile( $file )
 	{
-		self::open() -> log_file = $file; 
+		self::open()->log_file = $file;
 	}
 
 	/**
@@ -113,20 +117,20 @@ class Logger
 	 */
 	public static function setConfig( array $config )
 	{
-		self::open() -> config = $config + self::open() -> config;
+		self::open()->config = $config + self::open()->config;
 	}
 
 
 	/**
 	 * debug - Records debug information
-	 * 
+	 *
 	 * @param string $title 	The title of the log entry
 	 * @param mixed  $details	The log entry details
 	 */
-	public static function debug( $title, $details ) 
+	public static function debug( $title, $details )
 	{
-		self::open() -> record['debug'][][ $title ] = $details;
-		self::open() -> write( 'debug', $title, $details );
+		self::open()->record['debug'][][ $title ] = $details;
+		self::open()->write( 'debug', $title, $details );
 	}
 
 
@@ -136,10 +140,10 @@ class Logger
 	 * @param string $title 	The title of the log entry
 	 * @param mixed  $details	The log entry details
 	 */
-	public static function error( $title, $details ) 
+	public static function error( $title, $details )
 	{
-		self::open() -> record['error'][][ $title ] = $details;
-		self::open() -> write( 'error', $title, $details );
+		self::open()->record['error'][][ $title ] = $details;
+		self::open()->write( 'error', $title, $details );
 	}
 
 
@@ -149,10 +153,10 @@ class Logger
 	 * @param string $title 	The title of the log entry
 	 * @param mixed  $details	The log entry details
 	 */
-	public static function info( $title, $details ) 
+	public static function info( $title, $details )
 	{
-		self::open() -> record['info'][][ $title ] = $details;
-		self::open() -> write( 'info', $title, $details );
+		self::open()->record['info'][][ $title ] = $details;
+		self::open()->write( 'info', $title, $details );
 	}
 
 
@@ -162,10 +166,10 @@ class Logger
 	 * @param string $title 	The title of the log entry
 	 * @param mixed  $details	The log entry details
 	 */
-	public static function warn( $title, $details ) 
+	public static function warn( $title, $details )
 	{
-		self::open() -> record['warn'][][ $title ] = $details;
-		self::open() -> write( 'warn', $title, $details );
+		self::open()->record['warn'][][ $title ] = $details;
+		self::open()->write( 'warn', $title, $details );
 	}
 
 
@@ -176,23 +180,23 @@ class Logger
 	 * @param string $title 	The title of the log entry
 	 * @param mixed  $details	The log entry details
 	 */
-	public function write( $level, $title, $details ) 
+	public function write( $level, $title, $details )
 	{
 
-		if ( $this -> config[$level] )
+		if( $this->config[$level] )
 		{
 			if( is_array( $details ) )
 				$details = ArrayUtility::makeReadable( $details, true);
 
 			$timer_end = microtime(true);
-			$time = $timer_end - $this -> start_time;
+			$time = $timer_end - $this->start_time;
 
-			$line = date( 'G:i:s T D M j Y' )  . $this -> seperator . $time . $this -> seperator . $level . $this -> seperator;
-			$line .= $title . $this -> seperator . $details . "\n";
+			$line = date( 'G:i:s T D M j Y' )  . $this->seperator . $time . $this->seperator . $level . $this->seperator;
+			$line .= $title . $this->seperator . $details . "\n";
 
-			file_put_contents( $this -> log_file, $line, FILE_APPEND );
+			file_put_contents( $this->log_file, $line, FILE_APPEND );
 
-			$this -> lines_written++;
+			$this->lines_written++;
 		}
 	}
 
@@ -200,29 +204,34 @@ class Logger
 	/**
 	 * rotate - If log file size exceeds maximum, rotate.
 	 *
-	 * If the log size exceeds the maximum
+	 * If the log exists and its size exceeds the maximum
 	 * rename log by suffixing it with number.
-	 * If that file exists, rename it by incrementing that number.
+	 * If that file exists, rename it by incrementing that number until
+	 * the maximum number of old logs reached.
+	 * Starting with application.log.1
 	 */
 	public function rotate()
 	{
-		if ( file_exists ( $this -> log_file ) )
+		if( file_exists( $this->log_file ) )
 		{
-			$log_size = filesize( $this -> log_file );
+			$log_size = filesize( $this->log_file );
 
-			if ( $log_size > $this -> config['max_log_size'] )
+			if( $log_size > $this->config['max_log_size'] )
 			{
-				if( !file_exists( $this -> log_file . '.' . 1 ) )
-					rename( $this -> log_file, $this -> log_file . '.' . 1 );
+				if( !file_exists( $this->log_file . '.' . 1 ) )
+					rename( $this->log_file, $this->log_file . '.' . 1 );
 				else
 				{
 
-					for( $i =  $this -> config['max_old_logs']; $i > 0; $i-- )
+					for( $i =  $this->config['max_old_logs']; $i > 0; $i-- )
 					{
-						if( $i == $this -> config['max_old_logs'] )
-							unlink( $this -> log_file . '.' . $i );
-						else
-							rename( $this -> log_file . '.' . $i, $this -> log_file . '.' . ( $i + 1 ) );
+						if ( file_exists( $this->log_file . '.' . $i ) )
+						{
+							if( $i == $this->config['max_old_logs'] )
+								unlink( $this->log_file . '.' . $i );
+							else
+								rename( $this->log_file . '.' . $i, $this->log_file . '.' .( $i + 1 ) );
+						}
 					}
 				}
 			}
@@ -231,34 +240,37 @@ class Logger
 
 
 	/**
-	 * showDebug - Outputs debug data as a table 
+	 * showDebug - Outputs debug data as a table
 	 *
 	 * todo Displaying previous requests, session data, etc.
 	 */
 	public static function showDebug()
 	{
-		if( !self::open() -> config['display_debug'] ) 
+		if( !self::open()->config['display_debug'] )
 			return false;
 		else
 		{
 			$log_lines 	= null;
 
-			$log = fopen( self::open() -> log_file, "r");
-			while( ($line = fgetcsv($log, 0 , self::open()->seperator ) ) !== false ) 
-			{
-				$log_lines = "<tr>" . $log_lines;
+			$log = fopen( self::open()->log_file, "r");
 
-				$single_line = null;
-				foreach ($line as $cell) 
+			while(($line = fgetcsv($log, 0 , self::open()->seperator ) ) !== false )
+			{
+				$single_line = "<tr>";
+
+				foreach( $line as $cell )
 					$single_line .= "<td>" . htmlspecialchars($cell) . "</td>";
+
+				$single_line.= "</tr>\n";
+
 				$log_lines = $single_line . $log_lines;
-				
-				$log_lines.= "<tr>\n";
+
 			}
+
 			fclose($log);
 
 			$session 	= new Session( $autostart = false );
-			$session_dump 	= ArrayUtility::makeReadable( print_r( $session -> data, true) );
+			$session_dump 	= ArrayUtility::makeReadable( print_r( $session->data, true) );
 
 			if( !empty( $session_dump ) )
 				$session_dump = '<h2> Session Dump </h2><pre>' . $session_dump . '</pre><br/>';
@@ -285,6 +297,5 @@ TABLE;
 	}
 
 }
-
 
 ?>
